dashboard/requirements.txt
streamlit
pandas
matplotlib
python-dotenv
sqlalchemy
psycopg2-binary